Adding print statements for status updates.
First print statement denotes that a new crash file is ready to be
examined (reproduced) and is printed before the execution of the test.
> Processing crash file %s"

The second print statement reports successful submissions to the server.
Without that statement we wouldn't see hangs when submitting new crashes
to FuzzManager. The statement is printed once reproduction of the test
was successful and upload to the server was successful.
> Success: Submitted crash to server."