Implement #273 Signatures without a good testcase should show new crashes.
This adds a way of "watching" a signature to keep track of crash entries
submitted after the watch was added.

The user still needs to poll this new view, but it will help in cases where
the signature is interesting, but none of the crash entries found so far
will reproduce.