[EC2SpotManager] Redefine pool size as number of CPU cores. Previously it was instances.
This enables pools with mixed size instances. Price is now per core, so
the cheapest price per core will be chosen.

Size is treated as a minimum. eg. if pool size is one but the smallest
instance allowed has a core count of four, one instance will be created.

The data migration tries to do the right thing, but warnings should be
watched in case the calculated values are not desirable.

Fixes #494