[EC2SpotManager] Add pool update locking.
Only one cron task will run at a time, although sub-tasks are parallel.
The lock release is run as the final sub-task, so all related sub-tasks
are included in the mutex.

Use Redis instead of Fasteners, because we need an inter-process lock
that can be released by a different process than the one that acquired
it.