[SpotManager] Allow young instances to be unlisted.
There is an occasional race in the EC2 API between a spot request
being fulfilled, and the fulfilled instance being listed. To the
spotmanager algorithm, it looks like instances boot and then
immediately disappear, only to reappear a few minutes later (when we
have already removed it from our DB).

This introduces a 5 minute grace period, where instances can be
missing but remain in the DB. If they are still missing after this
time, they will be removed as before.