[taskmanager] Add logging and fix pool=null in some cases.
If a task worker type did not match an fuzzing-tc pool, we would
return early from the pool creation function, but continue to add
the task in the DB.

Fix this case, and add extra logging so cases like this are easier
to debug in the future.

Also add a management command to add a task group manually in case
existing tasks need to be added.