WIP: Add 'created' field to reportsummary and expose in UI (#939)
This only adds a _created_ field during cache generation, which
wouldn't make it work for old, cached reports.
It also involved quite a bit of vue hackery, due to the data
not being available during django page render.
Maybe the model ought to change to take not only a `collectionid`
but also a `collection.created`?
This is also WIP because I'm literally dumping the Python `created`
attribute into the JSON and reading it out in JS. There is no explicit
serialization/parsing of the date and I'm not sure what format
is going to be used. Heh.